// Neptune.Web/src/app/shared/generated/model/nearby-assets-result-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NearbyAssetDto } from './nearby-asset-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class NearbyAssetsResultDto { 
    RadiusMeters?: number;
    Assets?: Array<NearbyAssetDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface NearbyAssetsResultDtoForm { 
    RadiusMeters?: FormControl<number>;
    Assets?: FormControl<Array<NearbyAssetDto>>;
}

export class NearbyAssetsResultDtoFormControls { 
    public static RadiusMeters = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Assets = (value: FormControlState<Array<NearbyAssetDto>> | Array<NearbyAssetDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<NearbyAssetDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}